refactor(web): share CIDR and device-name validators
