+++ web/src/utils/index.ts
@@ -9,3 +9,4 @@ export * from './sorting';
 export * from './packetParser';
 export * from './clipboard';
 export * from './counterGroups';
+export * from './validation';
